Improve config file parsing.
Properly exit with an error and print if important conditions arise:

- Cannot read main config file
- Cannot read resources.cfg
- No log file defined in main config file

This was born out of the frustration of an accidental permisisons change
to resources.cfg and the pain required to track down why thousands of
Nagios checks suddenly started failing.

The poor sap who did a "restart" was confused when the system blew up
and nobody could identify what was happening until we dug into debug
logs and found that the macro parsing had gone awry.

Simply logging that resources.cfg is not readable is unacceptable. It
needs to be an obvious and showstopping error. Nagios should not attempt
to fly with one wing.

These changes will surely improve troubleshooting for admins everywhere.