openrc-init.in: add an OpenRC init script.
This commit adds an OpenRC init script that is integrated with
autotools for variable substitutions. The new init script is based on
the one used in Gentoo (also authored by yours truly), with a few
improvements from Github issue #404 and the autotools integration.

This init script should be usable out-of-the box, since the correct
paths to everything are provided by the user during ./configure.

This addresses Github issue #378.