Improve config file parsing. (#406)
* Improve config file parsing.

Properly exit with an error and print if important conditions arise:

- Cannot read main config file
- Cannot read resources.cfg
- No log file defined in main config file

This was born out of the frustration of an accidental permisisons change
to resources.cfg and the pain required to track down why thousands of
Nagios checks suddenly started failing.

The poor sap who did a "restart" was confused when the system blew up
and nobody could identify what was happening until we dug into debug
logs and found that the macro parsing had gone awry.

Simply logging that resources.cfg is not readable is unacceptable. It
needs to be an obvious and showstopping error. Nagios should not attempt
to fly with one wing.

These changes will surely improve troubleshooting for admins everywhere.

* Add missing newlines

My local patch to Nagios 3.x had these already. Missed when forward-ported.

* Add myself to list of contributors

* Update config.c

reverted most of the printfs to logits since they do printf when specified - but kept the log_file implementation

* Update config.c

remove the newlines from the updated logits

* Update config.c

`char * log_file` is specified in utils.c