openrc-init.in: reorder the prefix variables in the OpenRC init script.
In the OpenRC init script, the "prefix" and "exec_prefix" variables
were defined after the default value of NAGIOS_CONFIG was set to
"@sysconfdir@/nagios.cfg". If the user took the default value for
sysconfdir, that would end up as "${prefix}/etc/nagios.cfg", with
the prefix un-substituted (because the "prefix" variable is set
later). Obviously not what we want.

This commit moves the definitions of "prefix" and "exec_prefix" to the
very top, so that they can be used in NAGIOS_CONFIG, and fixes Github
issue #421.

Thanks to Troy Lea for noticing the problem.