Partly fix of issue: NagiosEnterprises/nagioscore#425. (#483)
* Partly fix of issue: NagiosEnterprises/nagioscore#425.
* Files: Changelog, THANKS updated.

This change updates the delay to not be from the very start of the program or the start of the first time a host/service goes NOT OK, but rather to be delayed from the last time it had a hard state change - which is where it would be expected.

e.g.: Under normal circumstances, the delay would only ever happen 1 time for the lifetime of the process. This introduces the delay to be present each time a notification is warranted.