update checks.c to use curly brackets with the same indentation as the conditional that begins them. probably didn't catch all of them. also adjusted function declarations to have curly brackets begin on the line after the declaration. will probably have a style guide soon of some kind.
the issue here was that when checking if a host or service was passive
some code was executed, but the code was a conditional. since everything
was only one execution, there were no brackets. it looked like this:

if (cr->check_type == CHECK_TYPE_PASSIVE)
	if (host_is_passive(hst, cr) == ERROR) {
		return ERROR;
		}
else
host_is_active(hst);

i wrote those codez, but only to keep in line with the current normal
core styling. hopefully you see the issue (the else was executed via
the nested if, not the appropriat conditional).

this was causing no more host checks to occur for this host once
it reached its max attempts.

this should fix quite a bit of issues that people have been seeing in
the master branch recently.

i also fixed the rest of the bracket styling and indentation so that
you can see where a logic chunk ends a bit easier. if anyone has any
input or feedback on the changes, i'm welcome to listen.

i'm almost positive this won't compile as is, but plan on fixing up
the corresponding header files with what i've removed tomorrow.

(bryan heden)