updated more in the t-tap directory. added a colorize.sh script i read that the tap harness has some colorization methods, but i'm not familiar with it at all - if someone reading this wants to try and get it done more naturally, that would be fine.
in my tests, the only two that work so far are test_logging and
test_checks. by "work" i mean actually execute.

so if you're following along at home, at the end of the `TESTS = ` group
of lines, overwrite it with the following:

`TESTS = "test_checks test_logging"`

then you can cd into nagioscore/t-tap and run the following:

```
make test
```

or

```
make each_test
```

i added the each_test target, because i like the verbose output. again,
there is probably a way to handle this with tap natively, but i'm not
aware. i'm more concerned with getting the testing to a level where it
is providing useful output (especially for the 4.4.0 release), and can
be relied upon for good data.