testing a better .travis.yml file
currently it runs make test, and THEN the coverage report.

well since the coverage report is always successful, the build is
always passing. this is less than ideal.

if anyone out there has any experience in dealing with this kind of thing
and there are some best-practices i'm ignoring, i'd love some help.