add flapping comment only if there is not already one
New flapping comments will only be added, if there is not already one.

This is necessary because Nagios will recheck the flapping state of all
hosts and services after reloading its configuration if large
installation tweaks are disabled.

Rechecking the flapping state resulted in a new comment added and the
old one was left in the comment list.