xdata/xsddefault.c: mark some global variables as "extern".
The pending release of gcc-10.x enables the -fno-common behavior by
default, which ultimately means that it will reject redefinitions of
global variables. Usually the fix is simple -- one of the variables
should be "extern" to declare it but to indicate that its definition
is found in another file.

In this case, the file xdata/xsddefault.c has defined a few variables
that are also defined in common/shared.c:

 * time_t program_start
 * int daemon_mode
 * time_t last_log_rotation
 * int enable_notifications
 * int execute_service_checks
 * int accept_passive_service_checks
 * int execute_host_checks
 * int accept_passive_host_checks
 * int enable_event_handlers
 * int obsess_over_services
 * int obsess_over_hosts
 * int enable_flap_detection
 * int process_performance_data
 * int nagios_pid

This commit marks them all "extern" in xdata/xsddefault.c, and afterwards
a "make all" succeeds with CFLAGS="-fno-common" under gcc-9.2.0.

Gentoo-bug: https://bugs.gentoo.org/706902
Issue: https://github.com/NagiosEnterprises/nagioscore/issues/721