Add @dougnazar's patch for #904. (resolves #904)
In the discussion of the issue, questions of performance were brought up. It looks to me like performance data commands are run exactly once per check, and that we also already do this operation at least once per scheduled check. Based on how much other code changed in and around 45681cb, I'm assuming this was an oversight.