configure.ac,Makefile.in: replace "which" with "command -v"
The "which" utility is commonly used to test for programs in the
user's $PATH. This has two downsides:

  1. Each invocation of "which" launches a new process, and this can
     cause minor performance issues.
  2. The "which" program need not be present on every system (it is
     certainly less common than, say, the "rm" program that we are
     using it to find).

This commit changes each "which" to "command -v" to avoid these two
issues. The "command" utility is standardized by POSIX,

  https://pubs.opengroup.org/onlinepubs/9799919799/utilities/command.html

and is typically implemented as a shell built-in rather than as a
separate utility.